♻️ refactor : Code Modularization #8 - improve the code structure by modularizing the code and separating functionalities.
